fix/configuration-capabilities-as-chip (#538)
